new data format working
